requirements.txt
# Core dependencies for Medical Imaging Pipeline Analysis
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.19.0
scipy>=1.7.0
pandas>=1.3.0

# GPU acceleration (optional)
cupy-cuda11x>=11.0.0; sys_platform != "darwin"
cucim>=22.0.0; sys_platform != "darwin"

# Medical imaging
monai>=1.0.0
nibabel>=3.2.0

# Utilities
psutil>=5.8.0
GPUtil>=1.4.0
networkx>=2.6.0
scikit-learn>=1.0.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0 